Lab6/requirements.txt
Django==3.1.2
django-heroku==0.3.1
gunicorn==20.0.4
social-auth-app-django
social-auth-core==3.3.3
python-social-auth
python-decouple==3.1
django-filter
drf-yasg==1.20.0
dj-database-url 
psycopg2
djangorestframework==3.12.2
django-rest-auth
django-allauth